전송할 정보의 수가 많으므로 여러번에 나눠서 보내야 함

+++ GLBlockJump/GLBlockServer/PacketStruct.h
@@ -29,11 +29,11 @@ struct PacketParam {
 
 
 struct PlayerInitInfo {
-    unsigned long long playerId;
+    int playerId;
     float spawnPos[3];
 };
 
 struct PktInitPlayers {
-    unsigned long long myPlayerId;
+    int myPlayerId;
     PlayerInitInfo players[3]; // 모든 플레이어의 초기 위치
 };
